Nowa wersja

+++ pegle/src/main/java/fantomit/zwalkowepegle/controllers/MainController.java
@@ -1,18 +1,26 @@
 package fantomit.zwalkowepegle.controllers;
 
+import android.content.Context;
 import android.util.Log;
 
+import com.annimon.stream.Collectors;
 import com.annimon.stream.Stream;
+import com.crashlytics.android.answers.Answers;
+import com.crashlytics.android.answers.CustomEvent;
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
 
 import java.util.ArrayList;
+import java.util.Calendar;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
+import java.util.concurrent.FutureTask;
 
 import de.greenrobot.event.EventBus;
 import fantomit.zwalkowepegle.APImodels.Station;
 import fantomit.zwalkowepegle.APImodels.StationListObject;
+import fantomit.zwalkowepegle.BuildConfig;
 import fantomit.zwalkowepegle.DBmodels.River;
 import fantomit.zwalkowepegle.DBmodels.Settings;
 import fantomit.zwalkowepegle.R;
@@ -21,16 +29,13 @@ import fantomit.zwalkowepegle.db.repositories.SettingsRepository;
 import fantomit.zwalkowepegle.db.repositories.StationRepository;
 import fantomit.zwalkowepegle.interfaces.MainActivityInterface;
 import fantomit.zwalkowepegle.utils.AktualizacjaEvent;
-import fantomit.zwalkowepegle.utils.LastDownloadEvent;
 import fantomit.zwalkowepegle.utils.RetroFitErrorHelper;
 import fantomit.zwalkowepegle.utils.UsuwanieRzekiEvent;
 import fantomit.zwalkowepegle.webservices.ListaStacjiWebService;
-import fantomit.zwalkowepegle.webservices.MyWebService;
 import fantomit.zwalkowepegle.webservices.StacjaWebService;
 import rx.Observable;
-import rx.Subscriber;
 import rx.android.schedulers.AndroidSchedulers;
-import rx.functions.Action1;
+import rx.schedulers.Schedulers;
 
 @Singleton
 public class MainController {
@@ -50,27 +55,38 @@ public class MainController {
     @Inject
     private EventBus eventBus;
 
+    public HashMap<String, Integer> plywalnoscRzek;
+
     private List<StationListObject> mListaStacji;
     private List<Station> mWojewodzkieStacje;
     private List<River> mRivers;
-    private Station mStacja;
     private int howManyStationsTested = 0;
     private boolean isDivided = false;
     public boolean isSorted = false;
+    private String mWojewodztwo = "";
 
     public void setView(MainActivityInterface mView) {
         this.mView = mView;
+        plywalnoscRzek = null;
         if (!eventBus.isRegistered(this)) {
             eventBus.register(this);
         }
         if (repoSettings.getSettings() == null) {
             repoSettings.createOrUpdate(new Settings());
         }
+        if(mView == null){
+            isSorted = false;
+        }
     }
 
     public void getListaStacji() {
         mView.showProgressSpinner();
         mRivers = repoRiver.getAll();
+        if (repoSettings.getSettings() == null) {
+            mView.displayToast("B³¹d bazy danych. Odinstaluj i zainstaluj aplikacjê ponownie");
+            return;
+        }
+        mWojewodztwo = repoSettings.getSettings().getWojewodztwo();
         if (mRivers == null || mRivers.isEmpty() || repoSettings.getSettings().isHasWojewodztwoChanged()) {
             mRivers.clear();
             repoRiver.deleteAll();
@@ -81,100 +97,84 @@ public class MainController {
             set.setHasWojewodztwoChanged(false);
             repoSettings.createOrUpdate(set);
 
-            result.observeOn(AndroidSchedulers.mainThread()).subscribe(new Action1<List<StationListObject>>() {
-                @Override
-                public void call(List<StationListObject> stationListObjects) {
-                    Log.i("Retrofit", "success");
-                    eventBus.post(new LastDownloadEvent(mView.getToday().toString()));
-                    mListaStacji = stationListObjects;
-                    if (mListaStacji != null && !mListaStacji.isEmpty()) {
-                        setWojewodztwo();
-                    } else {
-                        Log.e("FANTOM", "brak pobranych stacji");
-                    }
+            result.observeOn(AndroidSchedulers.mainThread()).subscribe(stationListObjects -> {
+                Log.i("Retrofit", "success");
+                mListaStacji = stationListObjects;
+                if (mListaStacji != null && !mListaStacji.isEmpty()) {
+                    setWojewodztwo();
+                } else {
+                    Log.e("FANTOM", "brak pobranych stacji");
                 }
-            }, new RetroFitErrorHelper(mView));
+            }, new RetroFitErrorHelper(null));
         } else {
             isDivided = true;
-            mView.displayRivers();
+            checkPlywalnosc();
         }
+    }
+
+    private void setWojewodztwo() {
+        if (BuildConfig.DEBUG)
+            Log.e("FANTOM", "Set " + repoSettings.getSettings().getWojewodztwo());
+        howManyStationsTested = 0;
+        if (mView != null)
+            mView.displayProgress(R.string.msg_setWojewodztwo, null);
+        Stream.of(mListaStacji)
+                .forEach((StationListObject s) -> getStacja(s.getId()));
+
+        if (BuildConfig.DEBUG)
+            Log.e("DOWNLOAD", "Downloaded " + Integer.toString(mListaStacji.size()) + " stations");
     }
 
     public void getStacja(String id) {
         Observable<Station> result = stacjaWS.getStacja(id);
         if (!isDivided) mWojewodzkieStacje = new ArrayList<>();
 
-        result.observeOn(AndroidSchedulers.mainThread()).subscribe(new Action1<Station>() {
-            @Override
-            public void call(Station station) {
-                mStacja = station;
-                if (mStacja != null && !isDivided) {
-                    //Log.e("FANTOM", "test stacji " + mStacja.getName());
-                    howManyStationsTested++;
-                    if (repoSettings.getSettings().getWojewodztwo().equals(mStacja.getStatus().getProvince())) {
-                        Log.e("SORTED", "Stacja " + mStacja.getName() + " dopasowana");
-                        if (mView != null)
-                            mView.displayProgress(0, "Stacja " + mStacja.getName() + " dopasowana");
-                        mWojewodzkieStacje.add(mStacja);
-                        if (repoStacja.findById(mStacja.getId()) != null) {
-                            Station s = repoStacja.findById(station.getId());
-                            station.setIsFav(s.isFav());
-                            station.setIsUserCustomized(s.isUserCustomized());
-                            station.setNotifByPrzeplyw(s.isNotifByPrzeplyw());
-                            station.setNotifCheckedId(s.getNotifCheckedId());
-                            if (s.getDolnaGranicaPrzeplywu() != -1.0) {
-                                station.setDolnaGranicaPrzeplywu(s.getDolnaGranicaPrzeplywu());
-                            } else {
-                                station.setDolnaGranicaPrzeplywu(s.getLw_przeplyw() != -1.0 ? s.getLw_przeplyw() : s.getLowDischargeValue());
-                            }
-
-                            if (s.getDolnaGranicaPoziomu() != -1) {
-                                station.setDolnaGranicaPoziomu(s.getDolnaGranicaPoziomu());
-                            } else {
-                                Double temp = new Double(s.getStatus().getLowValue());
-                                station.setDolnaGranicaPoziomu(s.getLw_poziom() != -1 ? s.getLw_poziom() : temp.intValue());
-                            }
-                        } else {
-                            Double temp = new Double(station.getStatus().getLowValue());
-                            station.setDolnaGranicaPoziomu(temp.intValue());
-                            station.setDolnaGranicaPrzeplywu(station.getLowDischargeValue());
+        result.observeOn(AndroidSchedulers.mainThread())
+                .subscribe(station -> {
+                    if (station != null && !isDivided) {
+                        howManyStationsTested++;
+                        if (mWojewodztwo.equals(station.getStatus().getProvince())) {
+                            Log.e("SORTED", "Stacja " + station.getName() + " dopasowana");
+                            if (mView != null)
+                                mView.displayProgress(0, "Stacja " + station.getName() + " dopasowana");
+                            saveStationToRepo(station);
                         }
-                        repoStacja.createOrUpdate(mStacja);
-                    }
-                    int progress = (howManyStationsTested / mListaStacji.size()) * 100;
-                    if (progress == 25) {
-                        Log.e("TESTED:", "25%");
-                    }
-                    if (progress == 50) {
-                        Log.e("TESTED:", "50%");
-                    }
-                    if (progress == 75) {
-                        Log.e("TESTED:", "75%");
-                    }
-                    if (progress == 100) {
-                        Log.e("TESTED:", "100%");
-                    }
 
-                    if (howManyStationsTested == mListaStacji.size()) {
-                        Log.e("TESTED", "Znaleziono: " + Integer.toString(mWojewodzkieStacje.size())
-                                + " stacji dla woj. " + repoSettings.getSettings().getWojewodztwo());
-                        sortRivers(mWojewodzkieStacje);
+                        if (howManyStationsTested == mListaStacji.size()) {
+                            sortRivers(mWojewodzkieStacje);
+                            if (BuildConfig.DEBUG)
+                                Log.e("TESTED", "Znaleziono: " + Integer.toString(mWojewodzkieStacje.size())
+                                        + " stacji dla woj. " + repoSettings.getSettings().getWojewodztwo());
+                        }
                     }
-                }
-            }
-        }, new RetroFitErrorHelper(mView));
+                }, new RetroFitErrorHelper(mView));
     }
 
-    private void setWojewodztwo() {
-        Log.e("FANTOM", "Set " + repoSettings.getSettings().getWojewodztwo());
-        howManyStationsTested = 0;
-        if (mView != null)
-            mView.displayProgress(R.string.msg_setWojewodztwo, null);
-        Stream.of(mListaStacji)
-                .forEach((StationListObject s) -> getStacja(s.getId()));
-        Log.e("DOWNLOAD", "Downloaded " + Integer.toString(mListaStacji.size()) + " stations");
-        if (mView != null)
-            mView.displayProgress(0, "Pobrano " + Integer.toString(mListaStacji.size()) + " stacji do posortowania");
+    private void saveStationToRepo(Station station){
+        mWojewodzkieStacje.add(station);
+        Station s = repoStacja.findById(station.getId());
+        if (s != null) {
+            station.setIsFav(s.isFav());
+            station.setIsUserCustomized(s.isUserCustomized());
+            station.setNotifByPrzeplyw(s.isNotifByPrzeplyw());
+            station.setNotifCheckedId(s.getNotifCheckedId());
+            if (s.getDolnaGranicaPrzeplywu() != -1.0) {
+                station.setDolnaGranicaPrzeplywu(s.getDolnaGranicaPrzeplywu());
+            } else {
+                station.setDolnaGranicaPrzeplywu(s.getLw_przeplyw() != -1.0 ? s.getLw_przeplyw() : s.getLowDischargeValue());
+            }
+            if (s.getDolnaGranicaPoziomu() != -1) {
+                station.setDolnaGranicaPoziomu(s.getDolnaGranicaPoziomu());
+            } else {
+                Double temp = new Double(s.getStatus().getLowValue());
+                station.setDolnaGranicaPoziomu(s.getLw_poziom() != -1 ? s.getLw_poziom() : temp.intValue());
+            }
+        } else {
+            Double temp = new Double(station.getStatus().getLowValue());
+            station.setDolnaGranicaPoziomu(temp.intValue());
+            station.setDolnaGranicaPrzeplywu(station.getLowDischargeValue());
+        }
+        repoStacja.createOrUpdate(station);
     }
 
     private void sortRivers(List<Station> stacje) {
@@ -182,40 +182,49 @@ public class MainController {
         if (mView != null)
             mView.displayProgress(R.string.msg_SortRivers, null);
         if (stacje.isEmpty()) Log.e("FANTOM", "brak stacji");
-        List<Station> customStations = new ArrayList<>();
         List<River> rzeki = new ArrayList<>();
         List<String> nazwyRzek = new ArrayList<>();
-        Stream.of(stacje)
-                .forEach(stacja -> {
-                    String riverName = stacja.getStatus().getRiver();
-                    if (!nazwyRzek.contains(riverName)) {
-                        River r = new River();
-                        r.setRiverName(riverName);
-                        r.addConnectedStation(stacja.getId());
-                        r.setTrend(stacja.getTrend());
-                        nazwyRzek.add(riverName);
-                        rzeki.add(r);
-                    } else {
-                        int i = 0;
-                        for (River r : rzeki) {
-                            if (r.getRiverName().equals(riverName)) {
-                                r.addConnectedStation(stacja.getId());
-                                r.setTrend(stacja.getTrend());
-                                rzeki.set(i, r);
+        FutureTask<List<River>> sorting = new FutureTask<>(() -> {
+            Stream.of(stacje)
+                    .forEach(stacja -> {
+                        String riverName = stacja.getStatus().getRiver();
+                        if (!nazwyRzek.contains(riverName)) {
+                            River r = new River();
+                            r.setRiverName(riverName);
+                            r.addConnectedStation(stacja.getId());
+                            r.setTrend(stacja.getTrend());
+                            nazwyRzek.add(riverName);
+                            rzeki.add(r);
+                        } else {
+                            int i = 0;
+                            for (River r : rzeki) {
+                                if (r.getRiverName().equals(riverName)) {
+                                    r.addConnectedStation(stacja.getId());
+                                    r.setTrend(stacja.getTrend());
+                                    rzeki.set(i, r);
+                                }
+                                i++;
                             }
-                            i++;
                         }
-                    }
-                });
-        nazwyRzek.clear();
-        if (!rzeki.isEmpty()) Log.e("FANTOM", "sortRivers-Succes");
-        Log.e("ILOSC RZEK w " + repoSettings.getSettings().getWojewodztwo(), Integer.toString(rzeki.size()));
-        mRivers = rzeki;
+                    });
+            return rzeki;
+        });
+        Observable<List<River>> observable = Observable.from(sorting);
 
-        mView.displayRivers();
-        for (River r : rzeki) {
-            repoRiver.createOrUpdate(r);
-        }
+        Schedulers.computation().createWorker().schedule(() -> sorting.run());
+
+        observable.subscribe((rivers) -> {
+            mRivers = rivers;
+            Stream.of(rivers).forEach(r -> repoRiver.createOrUpdate(r));
+            checkPlywalnosc();
+        });
+
+
+
+        nazwyRzek.clear();
+        if (BuildConfig.DEBUG) if (!rzeki.isEmpty()) Log.e("FANTOM", "sortRivers-Succes");
+        if (BuildConfig.DEBUG)
+            Log.e("ILOSC RZEK w " + repoSettings.getSettings().getWojewodztwo(), Integer.toString(rzeki.size()));
     }
 
 
@@ -225,8 +234,11 @@ public class MainController {
 
     public void onEvent(UsuwanieRzekiEvent event) {
         boolean czyUsunac = event.czyUsunac();
+        isSorted = false;
         if (czyUsunac) {
-            repoRiver.delete(mRivers.get(event.getRiverPos()));
+            for(int pos : event.getRiverPos()){
+                repoRiver.delete(mRivers.get(pos));
+            }
             mRivers = repoRiver.getAll();
             mView.displayRivers();
         }
@@ -246,30 +258,41 @@ public class MainController {
         }
     }
 
-    public HashMap checkPlywalnosc() {
-        Log.e("PLYWALNOSC", "STARTED");
-        HashMap<String, Integer> plywalnoscRzek = new HashMap<>();
-        List<StationListObject> listaStacji = listaWS.getListaStacji().toBlocking().first();
+    public void checkPlywalnosc() {
+        Log.e("Plywalnosc", "Rozpoczêto testowanie p³ywalnoœci");
+        long startTime = Calendar.getInstance().getTimeInMillis();
 
-        Stream.of(mRivers)
-                .forEach(river -> {
-                    List<String> connectedStations = river.getConnectedStations();
-                    int i = 0;
-                    for (StationListObject station : listaStacji) {
-                        if (connectedStations.contains(station.getId())) {
-                            if(repoStacja.findById(station.getId()) != null) {
-                                if (station.getPoziom() >= repoStacja.findById(station.getId()).getDolnaGranicaPoziomu()) {
-                                    i++;
-                                    plywalnoscRzek.put(river.getRiverId(), i);
-                                } else {
-                                    plywalnoscRzek.put(river.getRiverId(), i);
+        plywalnoscRzek = new HashMap<>();
+        Observable<List<StationListObject>> result = listaWS.getListaStacji();
+        result.observeOn(AndroidSchedulers.mainThread()).subscribe(listaStacji -> {
+                    Log.e("PLYWALNOSC", "data received");
+                    Stream.of(mRivers)
+                            .forEach(river -> {
+                                int i = 0;
+                                List<StationListObject> filteredList = Stream.of(listaStacji)
+                                        .filter(s -> river.getConnectedStations().contains(s.getId()))
+                                        .collect(Collectors.toList());
+                                for (StationListObject station : filteredList) {
+                                    Station s = repoStacja.findById(station.getId());
+                                    if (s != null) {
+                                        if (station.getPoziom() >= s.getDolnaGranicaPoziomu()) {
+                                            i++;
+                                            plywalnoscRzek.put(river.getRiverId(), i);
+                                            // Log.e("PLYWALNOSC", river.getRiverName());
+                                        } else {
+                                            plywalnoscRzek.put(river.getRiverId(), i);
+                                            //Log.e("PLYWALNOSC", river.getRiverName());
+                                        }
+                                    }
                                 }
-                            }
-                        }
-                    }
-                });
-        Log.e("PLYWALNOSC", "COMPLETED");
-        return plywalnoscRzek;
+                            });
+                    Answers.getInstance().logCustom(new CustomEvent("P³ywalnoœæ")
+                    .putCustomAttribute("TIME", (Calendar.getInstance().getTimeInMillis() - startTime)/1000));
+                    Log.e("PLYWALNOSC", "completed");
+                    mView.displayRivers();
+                }, new RetroFitErrorHelper(mView)
+        );
     }
-
 }
+
+
